find function replaces elements in state callback
Elements object was used for two purposes:
- specify the elements to take screenshot of
- specify the elements to act on.

In practice, this two sets almost never were identical,
so it makes sense to split them up.

As a first step, find function is introduced, which must
be used to find an elements to act on.

Also, all action now can receive selectors directly.