Replace codeclimate with coveralls
Codeclimate uses very outdated version of JSHint and has no intent of
updating it (I asked in May, the answer was "soon", but its October
now and its still not updated).
Amount of false positives on codeclimate now is far more then it will
be acceptable to be useful, so we are leaving it.
For coverage reports we will be using coveralls now, and for linting
`npm run lint` on Travis will suffice.