Add ability to ignore areas
Suite got new method `.ignoreElements`, that allows to set a number
of selectors, which should be ignored during tests.
After screenshot is taken, the ignored elements are filled with black
color on it before comparison.

The format of data, returned from `prepareScreenshot` JS call is changed
so that capture area and igonred area could be processed in a same way.