Restore configurable tolerance
Now it has a bit different meaning then before: it allows to set
max CIEDE2000 value not treated as difference. It is now also
configurable on a 3 levels: global(config), suite and state.

Due to possibility of different suite to have different tolerance
API method `buildDiff` is now deprecated: it has access only to global
tolerance, so diff it produces might not show exactly what caused test
to fail. Method is kept for backward compatibility, but users of a
programmatic API now encouraged to use `saveDiffTo` method of
test result  (passed to reporter) instead.