Allow to launch one WD session per suite
sessionMode option added which allows to switch between
perBrowser (1 session for browser, current behavior) and
perSuite (1 session for each suite).
To support both old and new behavior along with parallelLimit
setting, browser launcher was replaced with 3 pool classes.

New browser is requested from pool before each suite.

`Pool` class is the most simple variant of the pool which just opens
and closes session every time it requested.

`LimitedPool` wraps another pool and enforces the limit on the number
of currently open sessions. It is used when parallelLimit is set.

In perBrowser mode, limited or unlimited pool will be wrapped
in CachingPool which allows to share a session between multiple
suites.

As long as calibration is needed to be performed only once for
each browser id, it was moved to separate Calibrator class. This
allows to perform only once for each browser even in perSuite mode.