Session management options
Adds two new options:
- sessionsPerBrowser allows to run multiple parallel sessions at
once for the same browser id.
- suitesPerSession is a more flexible alternative to sessionMode
(which is now removed) that allows to limit number of suites run
in one browser session. With default Infinity value it is the same
as per browser session mode, with value of 1 it is the same as per
suite mode.